Change method index layout a bit to improve the search.
Now only the real method names are contents of the anchor element,
rather than both the name and the class. This way the JavaScript
search only searches the method names in the method index, avoiding
matching the classnames not only in the class index, but also in
the method index.