Add a "useful" default start page.
Previously it just said "This is the RDoc documentation." and didn’t
even display the navigation sidebar, making the docs unreadable. Now
it contains some helpful instructions to get a useful start page into
the docs, and also displays the sidebar so users can read the rest of
the documentation nevertheless.